+++ src/main/js/plugin/input-bindings/boolean.ts
@@ -10,7 +10,7 @@ import * as UiUtil from '../../controller/ui-util';
 import * as BooleanConverter from '../../converter/boolean';
 import {InputValue} from '../../model/input-value';
 import {ViewModel} from '../../model/view-model';
-import {InputBindingPlugin} from './input-binding-plugin';
+import {InputBindingPlugin} from '../input-binding';
 
 function createConstraint(params: InputParams): Constraint<boolean> {
 	const constraints: Constraint<boolean>[] = [];
