+++ src/main/js/plugin/input-bindings/color-object.ts
@@ -6,7 +6,7 @@ import {Color, RgbaColorObject, RgbColorObject} from '../../model/color';
 import {InputValue} from '../../model/input-value';
 import {ViewModel} from '../../model/view-model';
 import * as StringColorParser from '../../parser/string-color';
-import {InputBindingPlugin} from './input-binding-plugin';
+import {InputBindingPlugin} from '../input-binding';
 
 /**
  * @hidden
