+++ src/main/js/plugin/input-bindings/string.ts
@@ -11,7 +11,7 @@ import * as StringConverter from '../../converter/string';
 import {StringFormatter} from '../../formatter/string';
 import {InputValue} from '../../model/input-value';
 import {ViewModel} from '../../model/view-model';
-import {InputBindingPlugin} from './input-binding-plugin';
+import {InputBindingPlugin} from '../input-binding';
 
 function createConstraint(params: InputParams): Constraint<string> {
 	const constraints: Constraint<string>[] = [];
